+++ src/api/schema.ts
@@ -367,6 +367,10 @@ export const lists = {
       application: relationship({ ref: 'Application' }),
       agent: relationship({ ref: 'ExportContractAgent.exportContract' }),
       privateMarket: relationship({ ref: 'PrivateMarket.exportContract' }),
+      awardMethod: relationship({ ref: 'ExportContractAwardMethod' }),
+      otherAwardMethod: text({
+        db: { nativeType: 'VarChar(200)' },
+      }),
       finalDestinationKnown: nullableCheckbox(),
       finalDestinationCountryCode: text({
         db: { nativeType: 'VarChar(3)' },
@@ -436,6 +440,14 @@ export const lists = {
     },
     access: allowAll,
   },
+  ExportContractAwardMethod: list({
+    fields: {
+      value: text({
+        db: { nativeType: 'VarChar(50)' },
+      }),
+    },
+    access: allowAll,
+  }),
   PrivateMarket: list({
     fields: {
       exportContract: relationship({ ref: 'ExportContract.privateMarket' }),
@@ -773,6 +785,8 @@ export const lists = {
       paidByLetterOfCredit: checkbox(),
       totalContractValue: relationship({ ref: 'TotalContractValue' }),
       validExporterLocation: checkbox(),
+      isPartyToConsortium: checkbox(),
+      isMemberOfAGroup: checkbox(),
     },
     access: allowAll,
   }),